refactor(ivy): move instructions into separate directory
