+++ packages/core/src/render3/global_utils_api.ts
@@ -15,6 +15,6 @@
  * file in the public_api_guard test.
  */
 
-export {markDirty} from './instructions';
+export {markDirty} from './instructions/all';
 export {getPlayers} from './players';
 export {getComponent, getContext, getDirectives, getHostElement, getInjector, getListeners, getRootComponents, getViewComponent} from './util/discovery_utils';
